+++ source/stageSelect.cpp
@@ -72,7 +72,7 @@ void drawStageSelect()
 
 	if (stageNo)
 	{
-		const int stageX = (-40 * stageNo + screenWidth) / 2;
+		const int stageX = (-40 * static_cast<int>(stageNo) + screenWidth) / 2;
 
 		//Draw everything now
 		drawTexture(sprites[TEX_TEXTBOX], &rcCur[(stageSelectFlash >> 1) & 1], stageX + 40 * selectedStage, 64);
